pyproject.toml
[tool.poetry]
name = "governanceplatform"
version = "0.5.5+reporting-rc3"
description = "Governance Platform - NISINP"
authors = [
    "Cédric Bonhomme <cedric.bonhomme@nc3.lu>",
    "Jérôme Lombardi <jerome.lombardi@nc3.lu>",
    "Juan Rocha <juan.rocha@nc3.lu>",
    "Ruslan Baidan <ruslan.baidan@nc3.lu>",
]
license = "AGPL-3.0-or-later"
readme = "README.md"

homepage = "https://github.com/informed-governance-project/NISINP"
repository = "https://github.com/informed-governance-project/NISINP"
documentation = "https://nisinp.readthedocs.io"

keywords = ["NC3-LU", "NIS", "NIS2", "incident-notification"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.2"
psycopg2-binary = "^2.8.4"
cryptography = "^44.0.1"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.0"
drf-spectacular-sidecar = "^2024.1.1"
django-bootstrap5 = "^23.4"
django-otp = "^1.1.6"
django-formtools = "^2.5.1"
django-two-factor-auth = "^1.15.5"
phonenumbers = "^8.13.11"
django-import-export = "^3.2.0"
parler = "^1.0.1"
django-parler = "^2.3"
django-bootstrap-datepicker-plus = "^5.0.4"
django-countries = "^7.5.1"
weasyprint = "^65.1"
django-extensions = "^3.2.3"
django-parler-rest = "^2.2"
python-dateutil = "^2.8.2"
django-phonenumber-field = "^7.3.0"
django-filter = "^23.5"
pytz = "^2024.1"
plotly = "^5.23.0"
kaleido = "0.2.1"
openpyxl = "^3.1.5"
django-cookiebanner = "^0.2.8"
django-health-check = "^3.18.3"
whitenoise = "^6.9.0"
schedule = "^1.2.2"
django-simple-captcha = "^0.6.2"
celery = "^5.5.1"
redis = "^5.2.1"
gunicorn = "^23.0.0"
djangorestframework = "^3.16.0"
requests = "^2.32.4"
httpx = "^0.28.1"
djangorestframework-simplejwt = "^5.5.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
pre-commit = "^4.2.0"
black = "^25.1.0"
pyupgrade = "^3.19.1"
autoflake = "^2.3.1"
mypy = "^1.15.0"
types-requests = "^2.32.0.20250515"
pip-audit = "^2.9.0"
isort = "^6.0.1"
django-debug-toolbar = "^5.2.0"
pyparsing = "^3.2.3"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx-book-theme = "^1.0.1"
sphinxcontrib-openapi = "^0.8.1"
pydot = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true

show_error_context = true
pretty = true
